Support RM 5.1
