--- lib/patches/attachments_patch.rb
@@ -1,21 +1,0 @@
-require_dependency 'attachment'
-module RedmineLightbox2
-  module AttachmentsPatch
-    def self.included(base) # :nodoc:
-      base.class_eval do
-        unloadable # Send unloadable so it will not be unloaded in development
-
-        def download_inline
-          # apply before_filters
-          find_attachment
-          file_readable
-          read_authorize
-
-          send_file @attachment.diskfile, :filename => filename_for_content_disposition(@attachment.filename),
-                    :type => detect_content_type(@attachment),
-                    :disposition => 'inline'
-        end
-      end
-    end
-  end
-end
