+++ lib/redmine_lightbox2/hooks/view_layouts_base_html_head_hook.rb
